Display action plan sections depending on pia type
