+++ src/app/entry/sections/sections.component.ts
@@ -29,22 +29,7 @@ export class SectionsComponent implements OnInit {
   }
 
   async ngOnInit() {
-    const piaType = this._piaService.pia.type;
-    const dataNav = await this._appDataService.getDataNav();
-    let sections = {...dataNav.sections};
-
-    if(piaType == PiaType.regular) {
-      delete sections[3];
-      delete sections[2];
-    }
-
-    if(piaType === PiaType.simplified) {
-      delete sections[3];
-      delete sections[2];
-      delete sections[1];
-    }
-
-    this.data = {sections: Object.values(sections)};
+    this.data = await this._appDataService.getDataNav(this._piaService.pia);
 
     this.data.sections.forEach((section: any) => {
       section.items.forEach((item: any) => {
